implemented support for 64-bit bit-vector constants in Z3 and SMT2 backend, added test cases for unsigned long to test_result_wrapper and test_QF_BV

+++ src/metaSMT/result_wrapper.hpp
@@ -368,7 +368,7 @@ namespace metaSMT {
         : boost::logic::tribool(boost::logic::indeterminate))
         )
       { }
-      result_wrapper( unsigned value, unsigned width )
+      result_wrapper( unsigned long value, unsigned long width )
       : r( boost::dynamic_bitset<>(width, value) )
       { }
